restructur everything object related to objects directory
