+++ include/mathematic.h
@@ -2,16 +2,10 @@
 # define MATHEMATIC_H
 
 # include <math.h>
+# include "objects.h"
 # include "MLX42.h"
 
 typedef struct s_sphere	t_sphere;
-
-typedef struct s_vector
-{
-	double	x;
-	double	y;
-	double	z;
-}	t_vector;
 
 typedef struct s_ray
 {
@@ -29,9 +23,6 @@ t_vector		multiply_vec(t_vector vec, t_vector m);
 t_vector		mult_double_vec(double d, t_vector vec);
 t_vector		add_double_vec(double d, t_vector vec);
 t_vector		sub_double_vec(double d, t_vector vec);
-
-t_vector		*new_vec(double x, double y, double z);
-void			print_vec(t_vector vec);
 
 t_ray			*new_ray(t_vector origin, t_vector direction);
 uint32_t		ray_color(t_ray	*ray, t_sphere *sp);
