+++ src/main.c
@@ -1,16 +1,16 @@
 #include "minirt.h"
 
-int	main(int argc, char **argv)
+int	main(void)//int argc, char **argv)
 {
-	char	**file;
-	size_t	i;
+	// char	**file;
+	// size_t	i;
 
-	i = 0;
-	if (check_input(argc, argv))
-		return (EXIT_FAILURE);
-	file = get_file(argv[1]);
-	while (file[i] != NULL)
-		ft_printf("%s\n", file[i++]);
+	// i = 0;
+	// if (check_input(argc, argv))
+	// 	return (EXIT_FAILURE);
+	// file = get_file(argv[1]);
+	// while (file[i] != NULL)
+	// 	ft_printf("%s\n", file[i++]);
 	setup_window();
 	return (EXIT_SUCCESS);
 }
