+++ src/objects/vector.c
